Delete useless files HomeFragment, VideoFragment, MyPageFragment moved to fragments package
