+++ app/src/main/java/com/example/falldetectionapp/MainActivity.java
@@ -9,6 +9,9 @@ import androidx.fragment.app.FragmentTransaction;
 import android.os.Bundle;
 import android.view.MenuItem;
 
+import com.example.falldetectionapp.fragments.HomeFragment;
+import com.example.falldetectionapp.fragments.MyPageFragment;
+import com.example.falldetectionapp.fragments.VideoFragment;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.android.material.navigation.NavigationBarView;
