--- app/src/main/java/com/example/falldetectionapp/MyPageActivity.java
@@ -1,14 +1,0 @@
-package com.example.falldetectionapp;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class MyPageActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_my_page);
-    }
-}
