--- app/src/main/java/com/example/falldetectionapp/VideoActivity.java
@@ -1,14 +1,0 @@
-package com.example.falldetectionapp;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class VideoActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_video);
-    }
-}
