+++ app/src/main/java/com/example/falldetectionapp/fragments/VideoFragment.java
@@ -1,4 +1,4 @@
-package com.example.falldetectionapp;
+package com.example.falldetectionapp.fragments;
 
 import android.os.Bundle;
 
@@ -7,6 +7,8 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.example.falldetectionapp.R;
 
 /**
  * A simple {@link Fragment} subclass.
