--- app/src/main/res/layout/activity_my_page.xml
@@ -1,224 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical"
-    android:visibility="visible"
-    tools:context=".MyPageActivity">
-
-    <ImageButton
-        android:id="@+id/imageButton2"
-        android:layout_width="63dp"
-        android:layout_height="63dp"
-        android:layout_gravity="top|left"
-        app:srcCompat="?attr/actionModeCloseDrawable" />
-
-    <TextView
-        android:id="@+id/textView4"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:text="마이페이지"
-        android:textSize="30sp" />
-
-    <FrameLayout
-        android:layout_width="match_parent"
-        android:layout_height="100dp"
-        android:layout_weight="1"></FrameLayout>
-
-    <ImageView
-        android:id="@+id/imageView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        tools:layout_editor_absoluteX="141dp"
-        tools:layout_editor_absoluteY="105dp"
-        tools:srcCompat="@tools:sample/avatars" />
-
-    <FrameLayout
-        android:layout_width="match_parent"
-        android:layout_height="96dp"
-        android:layout_weight="1"></FrameLayout>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_gravity="center"
-        android:orientation="horizontal"
-        android:padding="20dp">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:paddingLeft="60dp"
-            android:paddingRight="10dp"
-            android:text="카메라 아이디"
-            android:textAlignment="center"
-            android:textSize="16sp"
-            android:textStyle="bold" />
-
-        <EditText
-            android:id="@+id/cameraIdEditText"
-            android:layout_width="145dp"
-            android:layout_height="match_parent"
-            android:layout_gravity="center"
-            android:ems="10"
-            android:inputType="textPostalAddress"
-            android:paddingLeft="10dp"
-            android:paddingRight="60dp" />
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="346dp"
-        android:layout_height="match_parent"
-        android:layout_gravity="center"
-        android:orientation="horizontal"
-        android:padding="20dp">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:paddingLeft="60dp"
-            android:paddingRight="10dp"
-            android:text="비밀번호"
-            android:textAlignment="center"
-            android:textSize="16sp"
-            android:textStyle="bold" />
-
-        <EditText
-            android:id="@+id/passwordEditText"
-            android:layout_width="145dp"
-            android:layout_height="match_parent"
-            android:layout_gravity="center"
-            android:ems="10"
-            android:inputType="textPostalAddress"
-            android:paddingLeft="10dp"
-            android:paddingRight="60dp" />
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="match_parent"
-        android:layout_gravity="center"
-        android:orientation="horizontal"
-        android:padding="20dp">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:paddingLeft="60dp"
-            android:paddingRight="10dp"
-            android:text="주소"
-            android:textAlignment="center"
-            android:textSize="16sp"
-            android:textStyle="bold" />
-
-        <EditText
-            android:id="@+id/addressEditText"
-            android:layout_width="145dp"
-            android:layout_height="match_parent"
-            android:layout_gravity="center"
-            android:ems="10"
-            android:inputType="textPostalAddress"
-            android:paddingLeft="10dp"
-            android:paddingRight="60dp" />
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="401dp"
-        android:layout_height="match_parent"
-        android:layout_gravity="center"
-        android:orientation="horizontal"
-        android:padding="20dp">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:paddingLeft="60dp"
-            android:paddingRight="10dp"
-            android:text="비상연락처 1"
-            android:textAlignment="center"
-            android:textSize="16sp"
-            android:textStyle="bold" />
-
-        <EditText
-            android:id="@+id/nokPhone_1_EditText"
-            android:layout_width="145dp"
-            android:layout_height="match_parent"
-            android:layout_gravity="center"
-            android:ems="10"
-            android:inputType="textPostalAddress"
-            android:paddingLeft="10dp"
-            android:paddingRight="60dp" />
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="403dp"
-        android:layout_height="match_parent"
-        android:layout_gravity="center"
-        android:orientation="horizontal"
-        android:padding="20dp">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:paddingLeft="60dp"
-            android:paddingRight="10dp"
-            android:text="비상연락처 2"
-            android:textAlignment="center"
-            android:textSize="16sp"
-            android:textStyle="bold" />
-
-        <EditText
-            android:id="@+id/nokPhone_2_EditText"
-            android:layout_width="145dp"
-            android:layout_height="match_parent"
-            android:layout_gravity="center"
-            android:ems="10"
-            android:inputType="textPostalAddress"
-            android:paddingLeft="10dp"
-            android:paddingRight="60dp" />
-
-    </LinearLayout>
-
-    <TextView
-        android:id="@+id/textView"
-        android:layout_width="107dp"
-        android:layout_height="wrap_content"
-        android:layout_gravity="top"
-        android:text="탈퇴하기" />
-
-    <TextView
-        android:id="@+id/textView3"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_gravity="top"
-        android:text="개인정보수정" />
-
-    <FrameLayout
-        android:layout_width="match_parent"
-        android:layout_height="124dp"
-        android:layout_weight="1"></FrameLayout>
-
-    <com.google.android.material.bottomnavigation.BottomNavigationView
-        android:id="@+id/bottomNavigationView"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom"
-        app:itemBackground="@color/white"
-        app:labelVisibilityMode="labeled"
-        app:menu="@menu/bottom_menu" />
-
-
-</LinearLayout>
