+++ app/src/main/res/layout/fragment_video.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".VideoFragment">
+    tools:context=".fragments.VideoFragment">
 
     <LinearLayout
         xmlns:app="http://schemas.android.com/apk/res-auto"
